Add recognized text to note content
This commit adds the ability to add recognized text from the OCR Capture Activity to the note content in the Note Detail Activity.

The `OcrCaptureActivity` has been updated to remove the toast messages for canceled and completed text recognition.

The `NoteDetailActivity` now checks for an "ocr_text" extra in the intent and sets the `editTextContent` to its value if it exists.

A new `.idea/vcs.xml` file has also been added to configure Git for the project.